grouping of threats by percentage of hp after battle some bug fixes

+++ core/src/main/java/compf/core/engine/pokemon/moves/Schedule.java
@@ -28,7 +28,7 @@ public class Schedule {
 		PokemonBattle _battle;
 		int _attacker;
 		int _defender;
-		byte _priority;
+		int _priority;
 		int _round;
 		int _id;
 		int _uid;
