+++ teamBuilder/src/main/kotlin/teambuilder/ImaginaryMove.kt
@@ -2,6 +2,6 @@
 package teambuilder
 import compf.core.engine.pokemon.moves.Move
 import compf.core.engine.pokemon.Type
-public class ImaginaryMove(power:Byte,type:Type,kind:MoveKind):Move(-1,"",power,1,100,0,type,kind,Move.TargetType.any){
+public class ImaginaryMove(power:Int,type:Type,kind:MoveKind):Move(-1,"",power,1,100,0,type,kind,Move.TargetType.any){
     
 }
